+++ src/lib.rs
@@ -1,0 +1,6 @@
+#[macro_use]
+extern crate serde_derive;
+extern crate serde;
+extern crate serde_json;
+
+pub mod models;
