+++ src/models/mod.rs
@@ -1,0 +1,3 @@
+pub mod button;
+pub mod controller;
+pub mod server;
